Split GSC server helpers
