+++ frontend/server/seo/gsc/constants.ts
@@ -1,0 +1,6 @@
+export const SEARCH_CONSOLE_SCOPE = 'https://www.googleapis.com/auth/webmasters.readonly';
+export const TOKEN_ENDPOINT = 'https://oauth2.googleapis.com/token';
+export const SEARCH_ANALYTICS_ROW_LIMIT = 25_000;
+export const DEFAULT_CACHE_TTL_MS = 30 * 60 * 1000;
+export const DASHBOARD_CACHE_VERSION = 2;
+export const DASHBOARD_DB_CACHE_PREFIX = 'seo.gsc.dashboard';
